requirements.txt
django>=5.0,<7
djangorestframework>=3.15
